baseconv: merge base32 and base64 into one util
