+++ Cargo.toml
@@ -93,8 +93,7 @@ generic = [
 redox = [
 
   # And maybe all generic utilities
-  "base32",
-  "base64",
+  "baseconv",
   "basename",
   "cksum",
   "comm",
@@ -147,8 +146,7 @@ default = ["generic", "unix"]
 [dependencies]
 uucore   = { path="src/uucore" }
 arch     = { optional=true, path="src/arch" }
-base32   = { optional=true, path="src/base32" }
-base64   = { optional=true, path="src/base64" }
+baseconv = { optional=true, path="src/baseconv" }
 basename = { optional=true, path="src/basename" }
 cat      = { optional=true, path="src/cat" }
 chgrp    = { optional=true, path="src/chgrp" }
