--- src/base64/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "base64"
-version = "0.0.1"
-authors = []
-build = "../../mkmain.rs"
-
-[lib]
-name = "uu_base64"
-path = "base64.rs"
-
-[dependencies]
-uucore = { path="../uucore" }
-
-[[bin]]
-name = "base64"
-path = "../../uumain.rs"
